requirements.txt
streamlit
torch
torchvision
numpy
Pillow
scikit-image
lime
grad-cam
opencv-python